+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.0.2
 
+* New logger `TagLogger` and now extension `Any#logger` will use it to decrease objects allocations
+
 ## 0.0.1
 
 * Project has been inited